Correct animation settings
