+++ app/src/main/java/app/simone/multiplayer/model/MatchUserInfo.kt
@@ -1,7 +1,5 @@
 package app.simone.multiplayer.model
 
-import app.simone.singleplayer.model.SColor
-
 /**
  * Created by nicola on 09/08/2017.
  */
