+++ app/src/main/java/app/simone/multiplayer/view/invites/PendingRequestsAdapter.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 import app.simone.R;
 import app.simone.multiplayer.model.FacebookUser;
 import app.simone.multiplayer.model.OnlineMatch;
-import app.simone.singleplayer.view.GameActivity;
+import app.simone.singleplayer.view.GameActivityImpl;
 
 import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
 
@@ -58,7 +58,7 @@ public class PendingRequestsAdapter extends ArrayAdapter<OnlineMatch> implements
 
             case R.id.item_info:
 
-                Intent intent = new Intent(mContext,GameActivity.class);
+                Intent intent = new Intent(mContext,GameActivityImpl.class);
                 intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
                 intent.putExtra("multiplayerMode", "multiplayerMode");
                 intent.putExtra("key",dataModel.getKey());
