+++ app/src/main/java/app/simone/singleplayer/messages/TimeToBlinkMsg.java
@@ -2,8 +2,9 @@ package app.simone.singleplayer.messages;
 
 import java.util.List;
 
-import app.simone.singleplayer.model.SColor;
+import app.simone.singleplayer.model.SimonColor;
 import app.simone.shared.messages.IMessage;
+import app.simone.singleplayer.model.SimonColorImpl;
 
 /**
  * TimeToBlinkMsg,
@@ -12,9 +13,9 @@ import app.simone.shared.messages.IMessage;
  */
 
 public class TimeToBlinkMsg implements IMessage {
-    private List<SColor> sequence;
+    private List<SimonColorImpl> sequence;
 
-    public TimeToBlinkMsg(List<SColor> sequence){
+    public TimeToBlinkMsg(List<SimonColorImpl> sequence){
         this.sequence = sequence;
     }
 
@@ -23,7 +24,7 @@ public class TimeToBlinkMsg implements IMessage {
         return MessageType.TIME_TO_BLINK_MSG;
     }
 
-    public List<SColor> getSequence(){
+    public List<SimonColorImpl> getSequence(){
         return this.sequence;
     }
 }
