+++ app/src/main/java/app/simone/singleplayer/model/SimonColor.java
@@ -1,0 +1,26 @@
+package app.simone.singleplayer.model;
+
+/**
+ * SimonColor interface
+ * @author Michele Sapignoli
+ */
+
+public interface SimonColor {
+    /**
+     * Gets the buttonId
+     * @return buttonId id of the button
+     */
+    int getButtonId();
+
+    /**
+     * Gets the soundId
+     * @return soundId id of the sound
+     */
+    int getSoundId();
+
+    /**
+     * Gets the colorId
+     * @return colorId id of the color
+     */
+    int getColorId();
+}
