+++ app/src/main/java/app/simone/singleplayer/model/SimonColorImpl.java
@@ -2,13 +2,13 @@ package app.simone.singleplayer.model;
 import app.simone.R;
 
 /**
- * SColor, enum.
+ * SimonColorImpl, enum.
  * Each value (GREEN, RED, BLUE, YELLOW) is associated with a button, a sound, and a color.
  *
  * @author Michele Sapignoli
  */
 
-public enum SColor {
+public enum SimonColorImpl implements SimonColor{
     GREEN(R.id.GREEN, R.raw.simonsound1, R.color.myGreen),
     RED(R.id.RED, R.raw.simonsound2, R.color.myRed),
     BLUE(R.id.BLUE, R.raw.simonsound3, R.color.myBlue),
@@ -18,39 +18,27 @@ public enum SColor {
     private final int soundId;
     private final int colorId;
 
-    SColor(final int buttonId, final int soundId, final int colorId) {
+    SimonColorImpl(final int buttonId, final int soundId, final int colorId) {
         this.buttonId = buttonId;
         this.soundId = soundId;
         this.colorId = colorId;
     }
 
-    /**
-     * Gets the buttonId
-     * @return buttonId id of the button
-     */
     public int getButtonId() { return buttonId; }
 
-    /**
-     * Gets the soundId
-     * @return soundId id of the sound
-     */
     public int getSoundId() { return soundId; }
 
-    /**
-     * Gets the colorId
-     * @return colorId id of the color
-     */
     public int getColorId() {
         return colorId;
     }
 
     /**
-     * Gets SColor from buttonId
+     * Gets SimonColorImpl from buttonId
      * @param number buttonId
-     * @return color SColor
+     * @return color SimonColorImpl
      */
-    public static SColor fromInt(int number){
-        for (SColor type : SColor.values()) {
+    public static SimonColorImpl fromInt(int number){
+        for (SimonColorImpl type : SimonColorImpl.values()) {
             if (type.getButtonId() == number) {
                 return type;
             }
@@ -64,7 +52,7 @@ public enum SColor {
      * @return colorId
      */
     public static int getColorIdFromButtonId(int buttonId){
-        for (SColor type : SColor.values()) {
+        for (SimonColorImpl type : SimonColorImpl.values()) {
             if (type.getButtonId() == buttonId) {
                 return type.colorId;
             }
