--- app/src/main/java/app/simone/singleplayer/view/IGameActivity.java
@@ -1,12 +1,0 @@
-package app.simone.singleplayer.view;
-
-import android.os.Handler;
-
-/**
- * Interface implemented by GameActivity.
- * @author Michele Sapignoli
- */
-
-public interface IGameActivity {
-    Handler getHandler();
-}
