+++ app/src/main/res/layout/activity_game.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:launchMode="singleTask"
-    tools:context="app.simone.singleplayer.view.GameActivity">
+    tools:context="app.simone.singleplayer.view.GameActivityImpl">
 
 
     <FrameLayout
