+++ app/src/main/res/layout/activity_settings.xml
@@ -27,7 +27,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:background="@color/myBlue"
+                android:background="@color/myRed"
                 android:orientation="vertical">
 
             </LinearLayout>
@@ -37,7 +37,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:background="@color/myBlue"
+                android:background="@color/myRed"
                 android:orientation="vertical">
 
 
@@ -51,7 +51,7 @@
             android:layout_width="match_parent"
             android:layout_height="fill_parent"
             android:layout_weight="1"
-            android:background="@color/myRed"
+            android:background="@color/myBlue"
             android:gravity="center_horizontal|center_vertical"
             android:orientation="horizontal"
             android:weightSum="2">
@@ -61,7 +61,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:background="@color/myRed"
+                android:background="@color/myBlue"
                 android:orientation="vertical">
 
             </LinearLayout>
@@ -71,7 +71,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:background="@color/myRed"
+                android:background="@color/myBlue"
                 android:orientation="vertical">
